Remove some redundancies and add api security into the api
