+++ backend/config/schema.py
@@ -1,15 +1,15 @@
 import graphene
-import graphql_jwt
+
 import products.schema
-
-class Query(products.schema.BrandQuery, products.schema.CategoryQuery, graphene.ObjectType):
-    title = graphene.String(default_value="Easy Commerce!")
+import users.schema
 
 
-class Mutation(products.schema.Mutation, graphene.ObjectType):
-    token_auth = graphql_jwt.ObtainJSONWebToken.Field()
-    verify_token = graphql_jwt.Verify.Field()
-    refresh_token = graphql_jwt.Refresh.Field()
+class Query(products.schema.Query, users.schema.Query):
+    pass
+
+
+class Mutation(products.schema.Mutation, users.schema.Mutation):
+    pass
 
 
 schema = graphene.Schema(mutation=Mutation, query=Query)
