+++ backend/products/types.py
@@ -1,0 +1,13 @@
+from graphene_django import DjangoObjectType
+
+from .models import Brand, Category
+
+
+class BrandType(DjangoObjectType):
+    class Meta:
+        model = Brand
+
+
+class CategoryType(DjangoObjectType):
+    class Meta:
+        model = Category
