+++ backend/users/mutations.py
@@ -1,0 +1,12 @@
+import graphene
+import graphql_jwt
+
+from .types import UserType
+
+
+class ObtainJSONWebToken(graphql_jwt.JSONWebTokenMutation):
+    user = graphene.Field(UserType)
+
+    @classmethod
+    def resolve(cls, root, info, **kwargs):
+        return cls(user=info.context.user)
