+++ backend/users/types.py
@@ -1,0 +1,18 @@
+from graphene_django import DjangoObjectType
+
+from users.models import User
+
+
+class UserType(DjangoObjectType):
+    class Meta:
+        model = User
+        fields = (
+            'id',
+            'username',
+            'email',
+            'first_name',
+            'last_name',
+            'is_superuser',
+            'is_staff',
+            'is_active',
+        )
